LayoutTests/inspector/formatting/is-text-likely-minified-expected.txt

Test minification detection.


== Running test suite: Formatting.isLikelyMinified
-- Running test case: JavaScriptShortSnippet.Unminified
PASS: This small JavaScript snippet should be classified as unminified.

-- Running test case: JQuerySnippet.Unminified
PASS: jQuery snippet should be classified as unminified.

-- Running test case: JQuerySnippet.Minified
PASS: jQuery snippet should be classified as minified.

-- Running test case: JQuerySnippet.MinifiedWithHeader
PASS: Source with leading unminified but trailing minified source should be classified as minified.
